+++ src/crantpy/queries/__init__.py
@@ -20,8 +20,13 @@ __getattr__, __dir__, __all__ = lazy_loader.attach(
         ],
         "nested_connectivity_matrices": [
             "All",
+            "ColumnOrderRule",
+            "DEFAULT_ORDER",
+            "DEFAULT_WITHIN_TYPE_ORDER",
             "DirectedNestedMatrix",
+            "EB_COLUMN_ORDER",
             "NestedMatrix",
+            "NeuronOrder",
             "NeuropilCollection",
         ],
         "neurons": [
@@ -43,9 +48,14 @@ __getattr__, __dir__, __all__ = lazy_loader.attach(
 __all__ = [
     "All",
     "CAVE_ROW_LIMIT",
+    "ColumnOrderRule",
+    "DEFAULT_ORDER",
+    "DEFAULT_WITHIN_TYPE_ORDER",
     "DirectedNestedMatrix",
+    "EB_COLUMN_ORDER",
     "NestedMatrix",
     "NeuronCriteria",
+    "NeuronOrder",
     "NeuropilCollection",
     "attach_synapses",
     "connections",